UDP communication is clean and nice
