--- UDP_network/fancyclient.h
@@ -1,5 +1,0 @@
-#ifndef UDP_FANCY_H
-#define UDP_FANCY_H
-
-
-#endif  // UDP_FANCY_H
