--- UDP_network/server.c
@@ -1,80 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <string.h>
-#include <sys/types.h>
-#include <sys/socket.h>
-#include <arpa/inet.h>
-#include <netinet/in.h>
-
-#define PORT 8080
-#define BUFFER_SIZE 1024
-
-int recvFile(char* buf, int s)
-{
-    int i;
-    char ch;
-    for (i = 0; i < s; i++) {
-        ch = buf[i];
-        if (ch == EOF)
-            return 1;
-    }
-    return 0;
-}
-
-
-
-int main() {
-    int sockfd;
-    char buffer[BUFFER_SIZE];
-    struct sockaddr_in servaddr, cliaddr;
-
-    // Creating socket file descriptor
-    if ((sockfd = socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
-        perror("socket creation failed");
-        exit(EXIT_FAILURE);
-    }
-
-    memset(&servaddr, 0, sizeof(servaddr));
-    memset(&cliaddr, 0, sizeof(cliaddr));
-
-    // Filling server information
-    servaddr.sin_family = AF_INET; // IPv4
-    servaddr.sin_addr.s_addr = INADDR_ANY;
-    servaddr.sin_port = htons(PORT);
-
-    // Bind the socket with the server address
-    if (bind(sockfd, (const struct sockaddr *)&servaddr, sizeof(servaddr)) < 0) {
-        perror("bind failed");
-        exit(EXIT_FAILURE);
-    }
-
-    socklen_t len;
-
-    len = sizeof(cliaddr); 
-	char* filename = "example_file2.wav";
-	ssize_t n;
-	socklen_t addr_size;
-
-	// Creating a file.
-	FILE* fp = fp = fopen(filename, "wb");	
-	while ((n = recvfrom(sockfd, buffer, BUFFER_SIZE, 0, (struct sockaddr *)&cliaddr, &len))>0) {
-		// receive
-		fwrite(buffer, 1, n, fp);
-		
-		bzero(buffer, BUFFER_SIZE);
-		
-		//printf("received %d bytes\n",n);
-
-		// process
-
-			//sendto(sockfd, buffer, strlen(buffer), MSG_CONFIRM, (const struct sockaddr *)&cliaddr, len);
-		//	break;
-		//}
-    }
-	fclose(fp);
-	printf("File received!\n");
-		
-
-    return 0;
-}
