--- UDP_network/server.h
@@ -1,5 +1,0 @@
-#ifndef UDP_SERVER_H
-#define UDP_SERVER_H
-
-
-#endif  // UDP_SERVER_H
