--- networking/server/server.c
@@ -1,71 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <arpa/inet.h>
-#include <netinet/in.h>
-#include <sys/socket.h>
-
-#define PORT 12345
-#define BUFFER_SIZE 1024
-
-int main() {
-    int server_socket, client_socket;
-    struct sockaddr_in server_addr, client_addr;
-    socklen_t addr_len = sizeof(client_addr);
-    char buffer[BUFFER_SIZE];
-    FILE *received_file;
-
-    // Create a server socket
-    server_socket = socket(AF_INET, SOCK_STREAM, 0);
-    if (server_socket < 0) {
-        perror("Error creating socket");
-        exit(EXIT_FAILURE);
-    }
-
-    server_addr.sin_family = AF_INET;
-    server_addr.sin_port = htons(PORT);
-    server_addr.sin_addr.s_addr = INADDR_ANY;
-
-    // Bind the server socket to a specific address and port
-    if (bind(server_socket, (struct sockaddr *)&server_addr, sizeof(server_addr)) < 0) {
-        perror("Bind failed");
-        exit(EXIT_FAILURE);
-    }
-
-    // Listen for incoming connections
-    if (listen(server_socket, 1) < 0) {
-        perror("Listen failed");
-        exit(EXIT_FAILURE);
-    }
-
-    printf("Server listening on port %d...\n", PORT);
-
-    // Accept a connection from a client
-    client_socket = accept(server_socket, (struct sockaddr *)&client_addr, &addr_len);
-    if (client_socket < 0) {
-        perror("Accept failed");
-        exit(EXIT_FAILURE);
-    }
-
-    printf("Connection accepted from %s:%d\n", inet_ntoa(client_addr.sin_addr), ntohs(client_addr.sin_port));
-
-    // Receive the WAV file data from the client and save it to a file
-    received_file = fopen("received_file.wav", "wb");
-    if (!received_file) {
-        perror("Error opening file for writing");
-        exit(EXIT_FAILURE);
-    }
-
-    int bytes_received;
-    while ((bytes_received = recv(client_socket, buffer, BUFFER_SIZE, 0)) > 0) {
-        fwrite(buffer, 1, bytes_received, received_file);
-    }
-
-    fclose(received_file);
-    printf("File received and saved as received_file.wav\n");
-
-    close(client_socket);
-    close(server_socket);
-
-    return 0;
-}
